+++ src/utils/modes.js
@@ -1,0 +1,51 @@
+/**
+ * Modes in Fortnite
+ */
+
+module.exports = {
+  SOLO: {
+    top: [10, 25],
+    id: 'p2',
+    name: 'Solo'
+  },
+  DUO: {
+    top: [5, 12],
+    id: 'p10',
+    name: 'Duo'
+  },
+  SQUAD: {
+    top: [3, 6],
+    id: 'p9',
+    name: 'Squad'
+  },
+  SOLOS3: {
+    top: [10, 25],
+    id: 'prior_p2',
+    name: 'Season 3 Solo'
+  },
+  DUOS3: {
+    top: [5, 12],
+    id: 'prior_p10',
+    name: 'Season 3 Duo'
+  },
+  SQUADS3: {
+    top: [3, 6],
+    id: 'prior_p9',
+    name: 'Season 3 Squad'
+  },
+  SOLOS4: {
+    top: [10, 25],
+    id: 'curr_p2',
+    name: 'Season 4 Solo'
+  },
+  DUOS4: {
+    top: [5, 12],
+    id: 'curr_p10',
+    name: 'Season 4 Duo'
+  },
+  SQUADS4: {
+    top: [3, 6],
+    id: 'curr_p9',
+    name: 'Season 4 Squad'
+  }
+};