+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog for Fortnite Bot
 
-### Unreleased
+## [3.10.0] - 2018-06-12
+### Changed
+- Specify username in bot's response when user hasn't played a mode
+- Specify username in bot's response when user not found
+- Move errors and modes into separate files from constants
+- Put utility methods, constants, errors, and modes into utils directory
+
+### Fixed
 - Fix message for when user has no recent matches
-- Put username instead of the word 'user' for when user hasn't played a mode
 - Fix regex matching for commands
-- Removed case insensitive flag on regex matching because text is made lowercase beforehand
 
 ## [3.9.1] - 2018-06-05
 ### Changed